moved the fragments file outside the activity dir to the activity dir
